Adding UITests and finishing the project
