+++ EasyMoneyExchanger/ CoreData/Rates+CoreDataProperties.swift
@@ -19,7 +19,3 @@ extension Rates {
     @NSManaged public var timeStamp: Int64
 
 }
-//
-//extension Rates : Identifiable {
-//
-//}
