+++ EasyMoneyExchanger/EasyMoneyExchanger/Constants/ButtonType.swift
@@ -1,0 +1,12 @@
+//
+//  ButtonType.swift
+//  EasyMoneyExchanger
+//
+//  Created by Leon on 14/12/20.
+//
+
+import Foundation
+enum ButtonType {
+    case to
+    case from
+}
