+++ EasyMoneyExchanger/EasyMoneyExchanger/Coordinator/Storyboarded.swift
@@ -9,7 +9,7 @@ import Foundation
 import UIKit
 
 protocol Storyboarded {
-    static func instantiate() -> Self?
+    static func instantiate(from storyboardName: UIStoryboard.Name) -> Self?
 }
 
 extension Storyboarded where Self: UIViewController {
