+++ EasyMoneyExchanger/EasyMoneyExchanger/Resources/Images/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -2,6 +2,15 @@
   "colors" : [
     {
       "idiom" : "universal"
+    },
+    {
+      "appearances" : [
+        {
+          "appearance" : "luminosity",
+          "value" : "dark"
+        }
+      ],
+      "idiom" : "universal"
     }
   ],
   "info" : {
