+++ EasyMoneyExchanger/EasyMoneyExchanger/Screens/Utils/UpdateLabels.swift
