+++ EasyMoneyExchanger/EasyMoneyExchangerTests/DataConverterTests.swift
@@ -10,35 +10,45 @@ import XCTest
 
 class DataConverterTestsTests: XCTestCase {
 
-    let supportedCurrencies = SupportedCurrenciesViewModel(coreData: CoreDataManager())
+    private var viewModel: SupportedCurrenciesViewModel!
+
+    override func setUp() {
+        super.setUp()
+        viewModel =  SupportedCurrenciesViewModel(coreData: CoreDataManager())
+    }
+
+    override func tearDown() {
+        super.tearDown()
+        viewModel = nil
+    }
 
     func testGetSupportedList() {
         let dictionary = [ "USD": "United States Dollar", "BRL": "Brazilian Real"]
-        let result = supportedCurrencies.getSupportedList(supportedDictionary: dictionary)
+        let result = viewModel.getSupportedList(supportedDictionary: dictionary)
         XCTAssert(result[0].currencyCode == "BRL" && result[0].currencyName == "Brazilian Real", "Supported List Operation Failed" )
     }
 
     func testSortSupportedList() {
         let list = [SupportedList(currencyCode: "USD", currencyName: "United States Dollar"), SupportedList(currencyCode: "BRL", currencyName: "Brazilian Real")]
-        let result = supportedCurrencies.sortSupportedList(supportedList: list)
+        let result = viewModel.sortSupportedList(supportedList: list)
         XCTAssert(result[0].currencyCode == "BRL" && result[0].currencyName == "Brazilian Real", "Supported List Operation Failed" )
     }
 
     func testAddSupportedListFlags() {
         let list = [SupportedList(currencyCode: "USD", currencyName: "United States Dollar"), SupportedList(currencyCode: "BRL", currencyName: "Brazilian Real")]
-        let result = supportedCurrencies.addSupportedListFlags(supportedList: list)
+        let result = viewModel.addSupportedListFlags(supportedList: list)
         XCTAssert(result[0].currencyCode == "USD" && result[0].currencyName == "🇺🇸 United States Dollar", "Supported List Operation Failed" )
     }
 
     func testGetStringFirstCharacter() {
         let list = "Brazilian Real"
-        let result = supportedCurrencies.getStringFirstCharacter(string: list)
+        let result = viewModel.getStringFirstCharacter(string: list)
         XCTAssert(result == "B", "Supported List Operation Failed" )
     }
 
     func testGetSupportedTitles() {
         let list = [SupportedList(currencyCode: "USD", currencyName: "United States Dollar"), SupportedList(currencyCode: "BRL", currencyName: "Brazilian Real")]
-        let result = supportedCurrencies.getSupportedTitles(supportedList: list)
+        let result = viewModel.getSupportedTitles(supportedList: list)
         XCTAssert(result == ["B", "U"], "Supported List Operation Failed" )
     }
 }
